refactor(presence): separate registry from claim invariants
